remove some outdated tests
